fix: replace tsup with tsdown
Replace tsup with tsdown for ALS and MCP builds to keep dual dist/lib
output while moving off tsup's TS6 baseUrl deprecation (egoist/tsup#1388).
Manual tsdown configs preserve bundled production and unbundled dev profiles.

- Swap root devDependency from tsup to tsdown and refresh lockfile
- Add tsdown.config.ts for ansible-language-server and ansible-mcp-server
- Remove tsup.config.ts and update compile, build:dist, and watch scripts
- Point Taskfiles, dictionary, verify-local skill, and tests at tsdown
- Move MCP resource data copy from tsup onSuccess to tsdown build:done
- Exclude tsdown.config.ts from package tsconfigs and ESLint project service

Related: egoist/tsup#1388
