+++ .agents/skills/verify-local/SKILL.md
@@ -132,7 +132,7 @@ Do NOT pipe the command through other commands when capturing exit codes
 
 | # | Check | Command | Time | What it catches |
 |---|-------|---------|------|-----------------|
-| 1 | build | `task build` | ~30s | TypeScript errors, missing exports, tsup bundling |
+| 1 | build | `task build` | ~30s | TypeScript errors, missing exports, tsdown bundling |
 | 2 | package | `task package` | ~45s | Knip dead code, npm packaging, vitest list |
 | 3 | lint | `prek run --all-files` | ~30s | ESLint, biome, ruff, codespell, cspell, yamllint, shellcheck |
